closet-curator-app/requirements.txt
flask
torch
pillow       # For image processing, if needed
supabase